Fix debugger on macOS
We need to use lldb-mi (machine interface) adapter instead
of real lldb. lldb-mi is sipped with cpptools extension and thus
we should use it if there is no system one (which is a rare case).

Also, use "environment" key in dbg dict instead of "env" since
lldb doesn't support "env".