Support all values from 'command-echo-mode' enumeration
Previously, we only supported two values from the 'command-echo-mode'
property (they are 'command-line' and 'sumamry'). For this we used the
boolean 'showCommandLines' setting in the plugin.

But now we also use the 'silent' and the 'command-line-with-environment'
values. To do this, we renamed the 'showCommandLines' setting with the
'commandEchoMode' and made it an enumeration.

Task-number: #47