Add runtime dependency on 'jsonc-parser`
Previous commit c08f632 fixes parsing for the JSON files containing
comments using the additional `jsonc-parser` library. But that commit
does not adds this library to the runtime dependencies, which leads to
the inoperability of the released extension v1.0.1.

We need to add this library and to the runtime dependencies.