Unify configurarion-name and overriden-properties to one place
We need to use the `configurations.json` file which describes
the all build configurations with its overridden properties.

In this case we don't need in the `overridden-properties.json`
file, and in the `qbs.overriddenProperties` command.

Instead, we use the `qbs.editConfiguration` command, and
the `configurations.json` file which creates automatically
at the first extension startup. Also the user can modify this
file as needs to add an other custom build configurations with
its overridden properties.

Task-number: #52