Add customizable filter on list of build profiles
A new settings file qbs-profiles.json may list names
of qbs build profiles. If the files exists and is not
empty, the list of selectable profiles will be
filtered.

There are new commands to create and edit the settings
file and the picker contains items to apply or ignore
the filter.